Move web to pkg and remove metrics dependency
